fix(add forward api and code refactor for forward)

+++ lib/src/models/platformfile_model.dart
@@ -15,7 +15,7 @@ class IsmchPlatformFile {
 
   final Uint8List? bytes;
 
-  final double? size;
+  final int? size;
 
   final String? extension;
 
@@ -23,7 +23,7 @@ class IsmchPlatformFile {
     String? path,
     String? name,
     Uint8List? bytes,
-    double? size,
+    int? size,
     String? extension,
   }) =>
       IsmchPlatformFile(
